Drop unversioned App flavour
Remove BaseApp and unify all logic on a single App contract with
the code of BaseApp with VersionedApp.
