+++ packages/lib/src/index.js
@@ -18,8 +18,7 @@ const assertRevert = helpers.assertRevert
 
 // model objects
 import Proxy from './proxy/Proxy'
-import UnversionedApp from './app/UnversionedApp'
-import VersionedApp from './app/VersionedApp'
+import App from './app/App'
 import Package from './package/Package'
 import ImplementationDirectory from './directory/ImplementationDirectory'
 import FreezableImplementationDirectory from './directory/FreezableImplementationDirectory'
@@ -41,8 +40,7 @@ export {
   Logger,
   FileSystem,
   Contracts,
-  VersionedApp,
-  UnversionedApp,
+  App,
   ImplementationDirectory,
   FreezableImplementationDirectory,
   Package,
